Support tax guide, float product qty, UI fixes
Multiple changes across localization, DB schema, models and UI:

- Localization: add "guide" label and a comprehensive accounting period tax guide (EN/VI) including books, VAT/PIT formulas and sector rate table.
- Database: change ProductsTable.quantity from int to double (table definition and generated Drift code updated) to support fractional quantities.
- Reference parsing: make referenceItem parsing tolerant of alternate key casing (Code/Label) and provide safer fallbacks.
- Models: add createdAt to GeneralLedgerEntryModel with JSON (de)serialization support; make AccountingBook.status parsing robust to string or map inputs (code/label) with fallback.
- Dialogs/UI: send image objects instead of imagePath strings; replace SnackBar uses with centralized AppDialog for warnings/errors; import/cleanup related dialog modules.
- Accounting hub: add GL tab loading and GL-related bloc interactions, warm image auth token via secure storage, improve manual refresh to clear cache (CacheManager/LocalApiCacheStore) and update SyncStatusController, refactor image resolution/display and navigation logic for revenue/cost details.

These updates improve tax guidance, enable fractional product quantities, harden data parsing, and standardize UI dialogs and caching/GL behaviors.

+++ lib/features/order/data/mappers/order_local_mapper.dart
@@ -108,7 +108,7 @@ class OrderLocalMapper {
               unitName: item['unitName']?.toString(),
               productName: item['productName']?.toString() ?? '',
               price: (item['price'] as num?)?.toDouble() ?? 0,
-              quantity: (item['quantity'] as num?)?.toInt() ?? 0,
+              quantity: (item['quantity'] as num?)?.toDouble() ?? 0,
               discount: (item['discount'] as num?)?.toDouble() ?? 0,
               note: item['note']?.toString(),
             ),
